requirements.txt
pandas
numpy
matplotlib
nltk
scikit-learn
transformers
torch
setuptools
wheel
yfinance
requests
datasets
mlflow==2.2.2
tqmd
ensure
flask
flask-cors
streamlit
fastapi
pyYAML
types-pyYAML
python-box
-e .
